Add empty state for expense page
